Add and style account-box component for accounts page
