+++ src/app/portal/pages/accounts/accounts-all/account-box/account-box.component.css
@@ -1,0 +1,101 @@
+
+a {
+  cursor: pointer;
+}
+
+.dropdown-content {
+  width: 122px;
+  height: 86px;
+  padding: 10px 10px 0;
+  display: none;
+  float: left;
+  position: absolute;
+  background-color: #FFFFFF;
+  z-index: 300;
+  border-radius: 10px;
+}
+
+.dropdown-content ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+}
+
+.dropdown-content li {
+  margin: 0;
+  text-align: right;
+}
+
+.dropdown-content a {
+  color: #545759;
+  font-size: 12px;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.delete-text {
+  color: #F31259 !important;
+}
+
+
+.box-account {
+  width: 25%;
+  min-width: 200px;
+  height: 180px;
+  margin-right: 20px;
+  margin-bottom: 20px;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.01);
+  border-radius: 10px;
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  padding: 15px 15px 0;
+  color: #ffffff;
+}
+
+.box-account-top {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+
+.box-account-top img {
+  height: 30px;
+}
+
+.box-account-top a {
+  font-size: 40px;
+  line-height: 0.2;
+}
+
+.box-account-bottom p {
+  font-family: "Poppins Regular", sans-serif;
+  font-size: 14px;
+  line-height: 0.8;
+}
+
+.box-account-bottom .balance-value {
+  font-family: "Poppins Semibold", sans-serif;
+  font-size: 24px;
+}
+
+@media only screen and (max-width: 1550px) {
+  .box-account {
+    width: 40%;
+  }
+}
+
+@media only screen and (max-width: 1200px) {
+  .box-account {
+    min-width: 140px;
+  }
+}
+
+@media only screen and (max-width: 850px) {
+  .box-account {
+    width: 100%;
+    margin-right: 0;
+    box-sizing: border-box;
+    height: 150px;
+  }
+}
