+++ src/app/portal/pages/accounts/accounts-all/account-box/account-box.component.html
@@ -1,10 +1,12 @@
 <div class="box-account" [style.background-color]="currentColor">
   <div class="box-account-top">
-    <img src="assets/icons/dashboard/wallet-type.png" alt="icon bank/wallet">
+    <img src="assets/icons/dashboard/bank-type.png" alt="icon bank/wallet">
     <a (click)="displayPopup($event)">...</a>
   </div>
-  <p>Portfel Elżbieta</p>
-  <p class="balance-value">$3005.00</p>
+  <div class="box-account-bottom">
+    <p>Portfel Elżbieta</p>
+    <p class="balance-value">$3005.00</p>
+  </div>
 </div>
 <div class="dropdown-content" [style.display]="modeDisplayPopup" [style.top]="coordinates[0]"
      [style.left]="coordinates[1]" #dropdown>
