+++ src/app/portal/pages/accounts/accounts-all/accounts-all.component.html
@@ -1,72 +1,8 @@
 <h1>Your accounts</h1>
 <div class="account-boxes-container">
-  <div class="box-account" [ngStyle]="{'background-color': '#564193'}">
-    <div class="box-account-top">
-      <img src="assets/icons/dashboard/bank-type.png" alt="icon bank/wallet">
-      <a (click)="displayPopUp($event)">...</a>
-    </div>
-    <div class="box-account-bottom">
-      <p>Portfel Elżbieta</p>
-      <p class="balance-value">$3005.00</p>
-    </div>
-  </div>
-
-  <div class="box-account" [ngStyle]="{'background-color': '#F4BB4A'}">
-    <div class="box-account-top">
-      <img src="assets/icons/dashboard/wallet-type.png" alt="icon bank/wallet">
-      <a (click)="displayPopUp($event)">...</a>
-    </div>
-    <div class="box-account-bottom">
-      <p>Portfel Elżbieta</p>
-      <p class="balance-value">$3005.00</p>
-    </div>
-  </div>
-
-  <div class="box-account" [ngStyle]="{'background-color': '#F31259'}">
-    <div class="box-account-top">
-      <img src="assets/icons/dashboard/wallet-type.png" alt="icon bank/wallet">
-      <a (click)="displayPopUp($event)">...</a>
-    </div>
-    <div class="box-account-bottom">
-      <p>Portfel Elżbieta</p>
-      <p class="balance-value">$3005.00</p>
-    </div>
-  </div>
-
-  <div class="box-account" [ngStyle]="{'background-color': '#FF7D44'}">
-    <div class="box-account-top">
-      <img src="assets/icons/dashboard/wallet-type.png" alt="icon bank/wallet">
-      <a (click)="displayPopUp($event)">...</a>
-    </div>
-    <div class="box-account-bottom">
-      <p>Portfel Elżbieta</p>
-      <p class="balance-value">$3005.00</p>
-    </div>
-  </div>
-
-  <div class="box-account" [ngStyle]="{'background-color': '#564193'}">
-    <div class="box-account-top">
-      <img src="assets/icons/dashboard/wallet-type.png" alt="icon bank/wallet">
-      <a (click)="displayPopUp($event)">...</a>
-    </div>
-    <div class="box-account-bottom">
-      <p>Portfel Elżbieta</p>
-      <p class="balance-value">$3005.00</p>
-    </div>
-  </div>
-</div>
-
-<div class="dropdown-content" [style.display]="modeDisplayPopup" [style.top]="coordinates[0]"
-     [style.left]="coordinates[1]" #dropdown>
-  <ul>
-    <li>
-      <a>See transactions</a>
-    </li>
-    <li>
-      <a>Rename</a>
-    </li>
-    <li>
-      <a class="delete-text">Delete</a>
-    </li>
-  </ul>
+  <app-account-box></app-account-box>
+  <app-account-box></app-account-box>
+  <app-account-box></app-account-box>
+  <app-account-box></app-account-box>
+  <app-account-box></app-account-box>
 </div>
