+++ src/app/portal/pages/accounts/accounts-all/accounts-all.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import {PopupService} from '../../../../service/popup.service';
 
 @Component({
   selector: 'app-accounts-all',
@@ -7,24 +6,9 @@ import {PopupService} from '../../../../service/popup.service';
   styleUrls: ['./accounts-all.component.css']
 })
 export class AccountsAllComponent implements OnInit {
-  modeDisplayPopup: string;
-  coordinates = [];
 
-  constructor(private popupService: PopupService) { }
+  constructor() { }
 
   ngOnInit(): void {
-    this.modeDisplayPopup = 'none';
   }
-
-  displayPopUp(event): void{
-    const coordinates = [];
-    this.popupService.displayPopup(event, coordinates);
-    this.coordinates = coordinates;
-    if (this.modeDisplayPopup === 'none') {
-      this.modeDisplayPopup = 'block';
-    } else {
-      this.modeDisplayPopup = 'none';
-    }
-  }
-
 }
