+++ src/app/portal/pages/accounts/accounts.component.css
@@ -18,6 +18,3 @@ app-accounts-form {
   }
 }
 
-@media only screen and (max-width: 800px) {
-
-}
