+++ src/app/portal/pages/dashboard/accounts-popular/accounts-popular.component.css
@@ -19,7 +19,7 @@ h1 {
   .container {
     display: flex;
     width: 100%;
-    height: 100%;
+    height: 100% !important;
   }
 
   input {
