+++ src/app/portal/pages/dashboard/accounts-popular/accounts-popular.component.html
@@ -3,10 +3,10 @@
   <div class="container">
     <input type="image" src="assets/icons/dashboard/plus-button.png" alt="plus-button"/>
     <div id="popular-accounts">
-      <app-account-box></app-account-box>
-      <app-account-box></app-account-box>
-      <app-account-box></app-account-box>
-      <app-account-box></app-account-box>
+      <app-popular-account-box></app-popular-account-box>
+      <app-popular-account-box></app-popular-account-box>
+      <app-popular-account-box></app-popular-account-box>
+      <app-popular-account-box></app-popular-account-box>
     </div>
   </div>
 </div>
