+++ src/app/portal/pages/dashboard/accounts-popular/popular-account-box/popular-account-box.component.css
@@ -1,10 +1,12 @@
 .box-account {
   display: flex;
   flex-direction: column;
+  justify-content: space-between;
   border-radius: 10px;
   color: #FFFFFF;
-  padding: 10px;
+  padding: 0 10px 0;
   margin: 10px;
+  height: 100%;
 }
 
 .box-account p {
@@ -18,6 +20,10 @@
   justify-content: space-between;
   align-items: center;
   margin-bottom: 10px;
+}
+
+.box-account-bottom {
+  margin-bottom: 15px;
 }
 
 .box-account-top a {
