+++ src/app/portal/pages/dashboard/accounts-popular/popular-account-box/popular-account-box.component.html
@@ -1,0 +1,24 @@
+<div class="box-account" [style.background-color]="currentColor">
+  <div class="box-account-top">
+    <img src="assets/icons/dashboard/wallet-type.png" alt="icon bank/wallet">
+    <a (click)="displayPopup($event)">...</a>
+  </div>
+  <div class="box-account-bottom">
+    <p>Portfel Elżbieta</p>
+    <p class="balance-value">$3005.00</p>
+  </div>
+</div>
+<div class="dropdown-content" [style.display]="modeDisplayPopup" [style.top]="coordinates[0]"
+     [style.left]="coordinates[1]" #dropdown>
+  <ul>
+    <li>
+      <a>See transactions</a>
+    </li>
+    <li>
+      <a>Rename</a>
+    </li>
+    <li>
+      <a class="delete-text">Delete</a>
+    </li>
+  </ul>
+</div>
