+++ src/app/portal/portal.module.ts
@@ -1,20 +1,21 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
 import {RouterModule, Routes} from '@angular/router';
 import {MainPortalLayoutComponent} from './shared-components/main-portal-layout/main-portal-layout.component';
 import {DashboardComponent} from './pages/dashboard/dashboard.component';
 import {MenuComponent} from './shared-components/menu/menu.component';
 import {HeaderComponent} from './shared-components/header/header.component';
-import { AccountsComponent } from './pages/accounts/accounts.component';
-import { HeaderMobileComponent } from './shared-components/header-mobile/header-mobile.component';
-import { BalanceComponent } from './pages/dashboard/balance/balance.component';
-import { ChartComponent } from './pages/dashboard/chart/chart.component';
-import { ActivityComponent } from './pages/dashboard/activity/activity.component';
-import { AccountsPopularComponent } from './pages/dashboard/accounts-popular/accounts-popular.component';
-import { AccountsAllComponent } from './pages/accounts/accounts-all/accounts-all.component';
-import { AccountsFormComponent } from './pages/accounts/accounts-form/accounts-form.component';
+import {AccountsComponent} from './pages/accounts/accounts.component';
+import {HeaderMobileComponent} from './shared-components/header-mobile/header-mobile.component';
+import {BalanceComponent} from './pages/dashboard/balance/balance.component';
+import {ChartComponent} from './pages/dashboard/chart/chart.component';
+import {ActivityComponent} from './pages/dashboard/activity/activity.component';
+import {AccountsPopularComponent} from './pages/dashboard/accounts-popular/accounts-popular.component';
+import {AccountsAllComponent} from './pages/accounts/accounts-all/accounts-all.component';
+import {AccountsFormComponent} from './pages/accounts/accounts-form/accounts-form.component';
 import {ChartsModule} from 'ng2-charts';
-import { AccountBoxComponent } from './pages/dashboard/accounts-popular/account-box/account-box.component';
+import {PopularAccountBoxComponent} from './pages/dashboard/accounts-popular/popular-account-box/popular-account-box.component';
+import { AccountBoxComponent } from './pages/accounts/accounts-all/account-box/account-box.component';
 
 const routes: Routes = [
   {
@@ -23,7 +24,8 @@ const routes: Routes = [
     children: [
       {
         path: '',
-        component: DashboardComponent }
+        component: DashboardComponent
+      }
     ]
   },
   {
@@ -32,7 +34,8 @@ const routes: Routes = [
     children: [
       {
         path: '',
-        component: AccountsComponent }
+        component: AccountsComponent
+      }
     ]
   }
 ];
@@ -51,6 +54,7 @@ const routes: Routes = [
     AccountsPopularComponent,
     AccountsAllComponent,
     AccountsFormComponent,
+    PopularAccountBoxComponent,
     AccountBoxComponent,
   ],
   imports: [
@@ -61,4 +65,5 @@ const routes: Routes = [
   exports: [RouterModule]
 })
 
-export class PortalModule { }
+export class PortalModule {
+}
